ss-177: * ui page
