+++ frontend/src/bundles/landing/components/components.ts
@@ -18,8 +18,10 @@ export { WalletCard } from './wallet-card/wallet-card';
 export {
     Button,
     DoughnutChart,
+    Dropdown,
     Header,
     Input,
     LineChart,
     Link,
+    MultiDropdown,
 } from '~/bundles/common/components/components';
