+++ frontend/src/bundles/ui/components/components.ts
@@ -5,9 +5,11 @@ export {
     CardTotal,
     Chart,
     DoughnutChart,
+    Dropdown,
     Header,
     Input,
     LineChart,
+    MultiDropdown,
     RangeSlider,
     Tabs,
     UserSettingsTabs,
@@ -19,8 +21,10 @@ export { CardPart } from './card-part/card-part';
 export { ChartPart } from './chart-part/chart-part';
 export { CodeHighlight } from './code-highlight/code-highlight';
 export { DoughnutChartPart } from './doughnut-chart-part/doughnut-chart-part';
+export { DropdownPart } from './dropdown-part/dropdown-part';
 export { HeaderPart } from './header-part/header-part';
 export { LineChartPart } from './line-chart-part/line-chart-part';
+export { MultiDropdownPart } from './multi-dropdown-part/multi-dropdown-part';
 export { RangeSliderPart } from './range-slider-part/range-slider-part';
 export { TabsPart } from './tabs-part/tabs-part';
 export { UserSettingsTabsPart } from './user-settings-tabs-part/user-settings-tabs-part';
